requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.3
django-crispy-forms==2.3
django-filter==24.3
django-htmx==1.21.0
django-redis==5.4.0
django-template-partials==24.4
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
Markdown==3.7
psycopg2==2.9.10
python-decouple==3.8
redis==5.2.0
sqlparse==0.5.1
